Create question2.sql
Question 2
From the three tables in Appendix A
a. Show schema generation query
b. Show the SQL query for number of customers purchasing more than 5 books
c. Show the SQL query for a list of customers who never purchased anything
d. Show the SQL query for list of book purchased with the users